pyproject.toml
[tool.poetry]
name = "signature-detection"
version = "0.1.0"
description = "Signature detection YOLOS model"
authors = ["Mario DEFRANCE <defrancemario@gmail.com>"]
license = "Apache 2.0"
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = ">=3.11,<3.13"
ipykernel = "^6.29.5"
ipywidgets = "^8.1.5"
optuna = "^4.2.1"
black = "^25.1.0"
flake8 = "^7.1.2"
flake8-pyproject = "^1.2.3"
pylint = "^3.3.5"
isort = "^6.0.1"
transformers = "^4.52.4"
datasets = "^3.6.0"
accelerate = "^1.8.1"
pytorch-lightning = "^2.5.2"
dotenv = "^0.9.9"
pillow = "^11.2.1"
torch = {version = "^2.6.0", source = "pytorch-gpu"}
torchvision = {version = "^0.21.0", source = "pytorch-gpu"}
torchaudio = {version = "^2.6.0", source = "pytorch-gpu"}
scipy = "^1.16.0"
coco-eval = "^0.0.4"
tensorboard = "^2.19.0"

[[tool.poetry.source]]
name = "pytorch-gpu"
url = "https://download.pytorch.org/whl/cu118"
priority = "explicit"

[tool.black]
line-length = 100

[tool.isort]
profile = "black"
line_length = 100

[tool.flake8]
max-line-length = 100
extend-ignore = ["E203", "E704"]

[tool.pylint.messages_control]
disable = ["C0301", "C0103", "E0110", "W0621", "import-error"]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
